Update Imageform state object, add handleChange
Imageform state is now an object with arrays of urls and files
representing images the user has added to the form. The form can now
accept multiple files of type '.png, .jpg, .jpeg'.

State: {urls: [], files: []} where files is an array of images as
ArrayBuffers and urls is an array of unique Strings representing each
image.

handleChange adds new images to the state when a user adds images to
the form.